+++ Assets/Examples/Scripts/UMa/Views/CameraView.cs
@@ -2,6 +2,8 @@ using System;
 using SimplifyIoC.Mediations;
 using UnityEngine;
 using UnityEngine.Events;
+using UnityEngine.Serialization;
+
 //============================================================
 //支持中文，文件使用UTF-8编码
 //@author	JiphuTzu
@@ -13,9 +15,7 @@ using UnityEngine.Events;
 namespace UMa.Views
 {
     public class CameraView : View {
-        [Serializable]
-        public class ClickEvent:UnityEvent<Vector3>{}
-        public ClickEvent OnClicked;
+        public UnityEvent<Vector3> onClick;
         private Camera _camera;
         protected override void Start()
         {
@@ -26,7 +26,7 @@ namespace UMa.Views
             if(!Input.GetMouseButtonUp(0)) return;
             var ray = _camera.ScreenPointToRay(Input.mousePosition);
             if(Physics.Raycast(ray,out var hit)){
-                OnClicked.Invoke(hit.point);
+                onClick.Invoke(hit.point);
             }
         }
 	}