+++ Assets/Examples/Scripts/UMa/Views/RecordView.cs
@@ -1,6 +1,5 @@
 using SimplifyIoC.Mediations;
 using UMa.Signals;
-using UnityEngine;
 using UnityEngine.UI;
 //============================================================
 //支持中文，文件使用UTF-8编码
@@ -25,7 +24,6 @@ namespace UMa.Views
         [ListensTo(typeof(RecordChangedSignal))]
         private void OnChanged(bool add)
         {
-            Debug.Log("record changed... "+add);
             if (add) Add();
             else Minus();
         }