+++ Assets/SimplifyIoC/Runtime/SimplifyIoC/Utils/BindEventAttribute.cs
@@ -6,40 +6,21 @@ using UnityEngine.Scripting;
 
 namespace SimplifyIoC.Utils
 {
-    [AttributeUsage(AttributeTargets.Field|AttributeTargets.Property)]
+    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Method)]
     public class BindEventAttribute : PreserveAttribute
     {
         public readonly string eventName;
-        public readonly string methodName;
+        public readonly string targetName;
 
         /// <summary>
         /// bind field event to some method
         /// </summary>
         /// <param name="eventName">The event name to bind</param>
-        /// <param name="methodName">The method name to bind. If null，will bind the method named "on"+filedName</param>
-
-        public BindEventAttribute(string eventName,string methodName = null)
+        /// <param name="targetName">The target field/property/method name to bind. If null，will bind the method named "on"+filedName</param>
+        public BindEventAttribute(string eventName, string targetName = null)
         {
             this.eventName = eventName;
-            this.methodName = methodName;
-        }
-    }
-    [AttributeUsage(AttributeTargets.Method)]
-    public class BindToEventAttribute : PreserveAttribute
-    {
-        public readonly string eventName;
-        public readonly string fieldName;
-
-        /// <summary>
-        /// bind field event to some method
-        /// </summary>
-        /// <param name="eventName">the event name to bind</param>
-        /// <param name="fieldName">the field name to bind</param>
-
-        public BindToEventAttribute(string eventName,string fieldName)
-        {
-            this.eventName = eventName;
-            this.fieldName = fieldName;
+            this.targetName = targetName;
         }
     }
 
@@ -50,79 +31,90 @@ namespace SimplifyIoC.Utils
         PROPERTY = 1 << 1,
         METHOD = 1 << 2
     }
+
     public static class BindEventExtension
     {
-        public static void BindEvents(this Component target,BindUsage usage = BindUsage.FIELD | BindUsage.PROPERTY | BindUsage.METHOD)
+        [Obsolete(
+            "use target.AddFieldParser(this.GetEventFieldParser(usage)).AddPropertyParser(this.GetEventPropertyParser(usage)).AddMethodParser(target.GetEventMethodParser(usage)).ParseAttributes() instead")]
+        public static void BindEvents(this Component target,
+            BindUsage usage = BindUsage.FIELD | BindUsage.PROPERTY | BindUsage.METHOD)
         {
-            if((usage & BindUsage.FIELD) == BindUsage.FIELD) BindFields(target);
-            if((usage & BindUsage.PROPERTY) == BindUsage.PROPERTY) BindProperties(target);
-            if((usage & BindUsage.METHOD) == BindUsage.METHOD) BindMethods(target);
+            target.AddAttributeParser(target.GetEventFieldParser(usage))
+                .AddAttributeParser(target.GetEventPropertyParser(usage))
+                .AddAttributeParser(target.GetEventMethodParser(usage))
+                .ParseAttributes();
         }
 
-        private static void BindMethods(Component target)
+        public static Action<T, BindEventAttribute, MethodInfo, Type> GetEventMethodParser<T>(this T target,
+            BindUsage usage = BindUsage.METHOD)
         {
-            var type = target.GetType();
-            var methods = type.GetMethods(BindingFlags.Instance
-                                          | BindingFlags.Public | BindingFlags.NonPublic);
-            foreach (var method in methods)
-            {
-                var attribute = method.GetCustomAttribute<BindToEventAttribute>();
-                if (attribute == null
-                    || string.IsNullOrWhiteSpace(attribute.eventName)
-                    || string.IsNullOrWhiteSpace(attribute.fieldName)) continue;
-                var obj = type.GetField(attribute.fieldName)?.GetValue(target) 
-                             ?? type.GetProperty(attribute.fieldName)?.GetValue(target);
-                var ue = GetEvent(obj, attribute.eventName);
-                if (ue == null) continue;
-                AddListener(ue, target, method);
-            }
+            return (usage & BindUsage.METHOD) == BindUsage.METHOD
+                ? MethodParser
+                : null;
         }
 
-        private static void BindFields(Component target)
+        public static Action<T, BindEventAttribute, FieldInfo, Type> GetEventFieldParser<T>(this T target,
+            BindUsage usage = BindUsage.FIELD)
         {
-            var type = target.GetType();
-            var fields = type.GetFields(BindingFlags.Instance
-                                        | BindingFlags.Public | BindingFlags.NonPublic);
-            foreach (var field in fields)
-            {
-                var attribute = field.GetCustomAttribute<BindEventAttribute>();
-                if (attribute == null || string.IsNullOrWhiteSpace(attribute.eventName)) continue;
-                var mn = attribute.methodName;
-                if (string.IsNullOrWhiteSpace(mn)) mn = $"on{field.Name}";
-                var methodInfo = type.GetMethod(mn,
-                    BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.IgnoreCase);
-                if (methodInfo == null) continue;
-                var ue = GetEvent(field.GetValue(target), attribute.eventName);
-                if (ue == null) continue;
-                AddListener(ue, target, methodInfo);
-            }
+            return (usage & BindUsage.FIELD) == BindUsage.FIELD
+                ? FieldParser
+                : null;
         }
 
-        private static void BindProperties(Component target)
+        public static Action<T, BindEventAttribute, PropertyInfo, Type> GetEventPropertyParser<T>(this T target,
+            BindUsage usage = BindUsage.PROPERTY)
         {
-            var type = target.GetType();
-            var fields = type.GetProperties(BindingFlags.Instance
-                                            | BindingFlags.Public | BindingFlags.NonPublic);
-            foreach (var field in fields)
-            {
-                var attribute = field.GetCustomAttribute<BindEventAttribute>();
-                if (attribute == null || string.IsNullOrWhiteSpace(attribute.eventName)) continue;
-                var mn = attribute.methodName;
-                if (string.IsNullOrWhiteSpace(mn)) mn = $"on{field.Name}";
-                var methodInfo = type.GetMethod(mn,
-                    BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.IgnoreCase);
-                if (methodInfo == null) continue;
-                var ue = GetEvent(field.GetValue(target), attribute.eventName);
-                if (ue == null) continue;
-                AddListener(ue, target, methodInfo);
-            }
+            return (usage & BindUsage.PROPERTY) == BindUsage.PROPERTY
+                ? PropertyParser
+                : null;
         }
 
-        private static void AddListener(UnityEventBase ue, Component target, MethodInfo method)
+        private static void MethodParser<T>(T target, BindEventAttribute attribute, MethodInfo method,
+            Type targetType)
+        {
+            if (string.IsNullOrWhiteSpace(attribute.eventName)
+                || string.IsNullOrWhiteSpace(attribute.targetName)) return;
+            const BindingFlags flags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic |
+                                       BindingFlags.IgnoreCase;
+            var obj = targetType.GetField(attribute.targetName, flags)?.GetValue(target)
+                      ?? targetType.GetProperty(attribute.targetName, flags)?.GetValue(target);
+            var ue = GetEvent(obj, attribute.eventName);
+            if (ue == null) return;
+            AddListener(ue, target, method);
+        }
+
+        private static void FieldParser<T>(T target, BindEventAttribute attribute, FieldInfo field, Type targetType)
+        {
+            if (string.IsNullOrWhiteSpace(attribute.eventName)) return;
+            var mn = attribute.targetName;
+            if (string.IsNullOrWhiteSpace(mn)) mn = $"on{field.Name}";
+            var methodInfo = targetType.GetMethod(mn,
+                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.IgnoreCase);
+            if (methodInfo == null) return;
+            var ue = GetEvent(field.GetValue(target), attribute.eventName);
+            if (ue == null) return;
+            AddListener(ue, target, methodInfo);
+        }
+
+        private static void PropertyParser<T>(T target, BindEventAttribute attribute, PropertyInfo property,
+            Type targetType)
+        {
+            if (string.IsNullOrWhiteSpace(attribute.eventName)) return;
+            var mn = attribute.targetName;
+            if (string.IsNullOrWhiteSpace(mn)) mn = $"on{property.Name}";
+            var methodInfo = targetType.GetMethod(mn,
+                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.IgnoreCase);
+            if (methodInfo == null) return;
+            var ue = GetEvent(property.GetValue(target), attribute.eventName);
+            if (ue == null) return;
+            AddListener(ue, target, methodInfo);
+        }
+
+        private static void AddListener(UnityEventBase ue, object target, MethodInfo method)
         {
             var type = ue.GetType();
             var ual = type.GetMethod("AddListener", BindingFlags.Instance | BindingFlags.NonPublic);
-            ual?.Invoke(ue, new object[] { target, method });
+            ual?.Invoke(ue, new[] { target, method });
         }
 
         private static UnityEventBase GetEvent(object target, string name)